Renamed all mock headers to include the mock_ prefix
