+++ firmware/boron-ins-fsm/test/base.h
@@ -5,10 +5,10 @@
 #include <iostream>
 
 // Mocks
-#include "mocks/eeprom.h"
-#include "mocks/Particle.h"
-#include "mocks/System.h"
-#include "mocks/os_queue_t.h"
+#include "mocks/mock_eeprom.h"
+#include "mocks/mock_Particle.h"
+#include "mocks/mock_System.h"
+#include "mocks/mock_os_queue_t.h"
 #include "mocks/mock_ble.h"
 
 // Particle.h library include files
