+++ firmware/boron-ins-fsm/test/mocks/mock_Particle.h
