+++ firmware/boron-ins-fsm/test/mocks/mock_eeprom.h
